refactor: split onboard shortcut write render
