+++ crates/daemon/src/onboard_cli.rs
@@ -116,34 +116,26 @@ pub use self::guided_config_support::{
 mod onboard_cli_render;
 #[path = "onboard_review_render.rs"]
 mod onboard_review_render;
+#[path = "onboard_shortcut_write_render.rs"]
+mod shortcut_write_render_support;
 pub use self::onboard_cli_render::{
     append_escape_cancel_hint, render_api_key_env_selection_screen_lines,
-    render_api_key_env_selection_screen_lines_with_default,
-    render_continue_current_setup_screen_lines, render_continue_detected_setup_screen_lines,
-    render_current_setup_write_confirmation_screen_lines, render_default_choice_footer_line,
-    render_detected_setup_write_confirmation_screen_lines,
-    render_existing_config_write_screen_lines, render_model_selection_screen_lines,
-    render_model_selection_screen_lines_with_default, render_onboard_entry_screen_lines,
-    render_onboarding_risk_screen_lines, render_provider_selection_screen_lines,
+    render_api_key_env_selection_screen_lines_with_default, render_default_choice_footer_line,
+    render_model_selection_screen_lines, render_model_selection_screen_lines_with_default,
+    render_onboard_entry_screen_lines, render_provider_selection_screen_lines,
     render_system_prompt_selection_screen_lines,
     render_system_prompt_selection_screen_lines_with_default,
-    render_write_confirmation_screen_lines,
 };
 use self::onboard_cli_render::{
     prompt_onboard_entry_choice, render_api_key_env_selection_screen_lines_with_style,
-    render_existing_config_write_header_lines_with_style,
     render_model_selection_screen_lines_with_style, render_onboard_choice_screen,
-    render_onboard_entry_interactive_screen_lines_with_style,
-    render_onboard_shortcut_header_lines_with_style, render_prompt_with_default_text,
+    render_onboard_entry_interactive_screen_lines_with_style, render_prompt_with_default_text,
     render_provider_selection_header_lines, render_system_prompt_selection_screen_lines_with_style,
     render_web_search_credential_selection_screen_lines_with_style, screen_subtitle,
     tui_header_style,
 };
 #[cfg(test)]
-use self::onboard_cli_render::{
-    render_onboard_option_lines, render_onboard_option_prefix,
-    render_onboard_shortcut_screen_lines_with_style,
-};
+use self::onboard_cli_render::{render_onboard_option_lines, render_onboard_option_prefix};
 #[cfg(test)]
 use self::onboard_review_render::provider_matches_for_review;
 use self::onboard_review_render::{
@@ -154,6 +146,19 @@ pub use self::onboard_review_render::{
     render_current_setup_review_lines_with_guidance,
     render_detected_setup_review_lines_with_guidance, render_onboard_review_lines_with_guidance,
     summarize_prompt_addendum, summarize_prompt_mode, summarize_provider_credential,
+};
+#[cfg(test)]
+use self::shortcut_write_render_support::render_onboard_shortcut_screen_lines_with_style;
+pub use self::shortcut_write_render_support::{
+    render_continue_current_setup_screen_lines, render_continue_detected_setup_screen_lines,
+    render_current_setup_write_confirmation_screen_lines,
+    render_detected_setup_write_confirmation_screen_lines,
+    render_existing_config_write_screen_lines, render_onboarding_risk_screen_lines,
+    render_write_confirmation_screen_lines,
+};
+use self::shortcut_write_render_support::{
+    render_existing_config_write_header_lines_with_style,
+    render_onboard_shortcut_header_lines_with_style,
 };
 #[cfg(test)]
 use self::starting_point_render_support::render_starting_point_selection_header_lines_with_style;
