+++ crates/daemon/src/onboard_screen_specs.rs
@@ -1,8 +1,8 @@
 use super::*;
-use crate::onboard_cli::onboard_cli_render::render_shortcut_default_choice_footer_line;
 use crate::onboard_cli::onboard_review_render::{
     build_onboard_review_digest_display_lines, onboard_display_line,
 };
+use crate::onboard_cli::shortcut_write_render_support::render_shortcut_default_choice_footer_line;
 
 pub(super) fn render_onboarding_risk_screen_lines_with_style(
     width: usize,
